* ergoemacs-functions.el: added ergoemacs-beginning-of-line-or-block and ergoemacs-end-of-line-or-block
might be usedful to set Alt+h to ergoemacs-beginning-of-line-or-block. When pressed for the first time, move to beginning of line, press again, it functions as ergoemacs-backward-block. Similarly for the key for ergoemacs-backward-block.

i think this would be improve efficiency, because: normally, Alt+h moves to beginning of line. But then it's dead there. Alt+h is a repeating key. Could be of some more use. Also, we need another keyspot for ergoemacs-backward-block, which by default is Alt+Shift+u, kinda ackward. If they are merged to one, i think we have 2 advantages. Simiraly for end of line and next text block commands.

the possible disadvantage is that if you want to just move to previous block, you have to press the key twice. So, it's Alt+h Alt+h vs Alt+Shift+u. I think these 2 keys are kinda a tie for ease-of-press, and so other advantages rules.