Fix command-loop errors
The problem was that mapkeymap was looking for a prefix in a keymap
and assuming that the result was a list.  This will fail when invoking
'term', because it has a keymap with a lot of elements specified as

  #^^[3 0 term-send-raw-meta term-send-raw-meta ...

We now check explicitly for whether the result is a list.